FEATURE - Show message counts in ~stats
