+++ hourai/bot.py
@@ -1,11 +1,12 @@
 import aiohttp
 import asyncio
 import collections
+import enum
 import itertools
 import logging
 import pkgutil
-import traceback
 import sys
+import traceback
 from discord.ext import commands
 from . import config, actions
 from .db import storage
@@ -13,6 +14,21 @@ from .utils import fake, format
 from .context import HouraiContext
 
 log = logging.getLogger(__name__)
+
+
+class CounterKeys(enum.Enum):
+    MESSAGES_RECIEVED   = 0x100             # noqa: E221
+    MESSAGES_DELETED    = 0x101             # noqa: E221
+    MESSAGES_EDITED     = 0x102             # noqa: E221
+    MEMBERS_JOINED      = 0x200             # noqa: E221
+    MEMBERS_LEFT        = 0x201             # noqa: E221
+    MEMBERS_BANNED      = 0x202             # noqa: E221
+    MEMBERS_UNBANNED    = 0x203             # noqa: E221
+    MEMBERS_VERIFIED    = 0x204             # noqa: E221
+    MEMBERS_REJECTED    = 0x205             # noqa: E221
+
+    def __repr__(self):
+        return self.name
 
 
 class CogLoadError(Exception):
